Fix bug causing divide by zero error
The projection function in the code didn't handle div by zero errors. This was fixed to not cause errors, but now there's the issue of what to do when there is a div by zero. I simply returned the numerator if denom=0, however, this should be further considered when addressing the projection function issue #3 .